fix(lint): Resolve react-refresh warnings by splitting exports
- Move useTour hook to separate file to avoid mixing component and hook exports
- Remove TooltipProvider re-export from Tooltip.tsx
- Update barrel exports accordingly

Reduces ESLint warnings from 52 to 50 (max allowed)
